+++ .rubocop.yml
@@ -7,6 +7,7 @@ Metrics/ModuleLength:
 Metrics/BlockLength:
   Exclude:
     - "**/*_spec.rb"
+    - "fusuma.gemspec"
 
 Metrics/LineLength:
   Max: 100
