+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-05-29 16:18:19 +0900 using RuboCop version 0.70.0.
+# on 2019-07-19 00:41:39 +0900 using RuboCop version 0.73.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -18,5 +18,5 @@ Metrics/MethodLength:
 Style/NumericPredicate:
   Exclude:
     - 'spec/**/*'
-    - 'lib/fusuma/plugin/vectors/rotate_vector.rb'
-    - 'lib/fusuma/plugin/vectors/swipe_vector.rb'
+    - 'lib/fusuma/plugin/detectors/rotate_detector.rb'
+    - 'lib/fusuma/plugin/detectors/swipe_detector.rb'
