+++ spec/lib/plugin/executors/dummy_detector.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-require './lib/fusuma/plugin/vectors/vector.rb'
+require './lib/fusuma/plugin/detectors/detector.rb'
 
 module Fusuma
   module Plugin
-    module Vectors
-      class DummyVector < Vector
+    module Detectors
+      class DummyDetector < Detector
         def initialize(finger, direction)
           @finger = finger
           @direction = direction
